Change BigInteger creation from using Strings to use a constructor that might reuse BigInteger instances
Not sure what the deal was with this approach to create new BigInteger objects, but it is problematic with Truffle. As far as I can tell, this should still be ok. We don't change the value of any BigInteger instance. So, reuse should not really be a problem, I think…

[POTENTIALLY-PROBLEMATIC]

Signed-off-by: Stefan Marr <git@stefan-marr.de>