Added the first working type system.
For some reason, previous attempts always failed.

My current understanding is that we need the type system in order to map SOM types to Java natives, avoiding boxing. Especially to utilize the return-type specialization approach of Truffle.

Problematic is I think at the moment that the generated type system uses automatically generated asObject() and isObject() functions, and does override the once I provide.

In the literal nodes for Integer, String, and BigInteger, I needed to manually add an executeGeneric, which is more optimal than the generated one, but seems to point out problems with the overriding of asObject()/isObject().

[TODO-REPORT: investigate how the use of the type system works out eventually, report the problems with the overriding]

Signed-off-by: Stefan Marr <git@stefan-marr.de>