Added support for inner iterations and added initial benchmark configuration
Inner iterations can be used to increase the problem size of a benchmark between measurements.

The initial benchmark configuration is tuned for RPySOM with support for a JIT compiler. It leaves still headroom for improvements of the RPySOM implementation, however, these settings might be too high for any of the other implementations.

Signed-off-by: Stefan Marr <git@stefan-marr.de>