Revise Node design and use SimpleLanguage approach to polymorphic inline caches
The node design is now roughly based on the SimpleLanguage approach.
From there, the approach to Send nodes is taken. Send nodes are used for caching and inlining.
However, compared to the previous approach, this means that we have another level of indirection to the primitives.
The last refactoring made primitives direct nodes in the AST. Thus, the node for a specific send operation was a node that could specialize for the primitive operation. Now however, we layer before that the polymorphic inline cache. From a theoretical perspective, this might be cleaner, since the act of sending, and the operation remain decoupled, but it introduces another layer of indication. I think, I would have preferred if the TruffleDSL would support state and we could realize the inline cache as part of the specialization for primitive nodes directly. This would avoid another layer of indirections and make a monomorphic send node just another specialization, similar to the primitive specializations.
Also, I am not sure yet whether the new design allows for proper type specialization because of the way the send nodes work.

The hierarchy of nodes also changed again. Primitive nodes are not monomorphic types of nodes anymore, since this was refactored out to send nodes.
The `cloneForInlining()` is now realized based on a hack of the TruffleDSL to produce a polymorphic `createUninitialized()` function for each node.
This way, I can easily get a copy in an uninitialized state for each node in the AST.

Signed-off-by: Stefan Marr <git@stefan-marr.de>