Added `isAlwaysToBeInlined()` inspired by SimpleLanguage
This is meant for primitives and perhaps really small methods like accessors.
There are a couple of candidates methods in the boolean classes, Object, Nil, and HashEntry that return literals, self, or fields. These could benefit from this optimization.
And hopefully primitives are then the direct nodes directly behind a Send node to avoid to many indirections.

Signed-off-by: Stefan Marr <git@stefan-marr.de>