Fixed the implementation of `super`
Super was implemented incorrectly. `super` requires lexical semantics, and not dynamic semantics, which is now encoded directly in the node. However, since we do not have a reference to the relevant class before parsing is completed, we need to lazily update the AST. And, we use a Uninitialized node for that.

Signed-off-by: Stefan Marr <git@stefan-marr.de>