Removed InitializeTemporarySlotsNode again, replaced by special case of LocalVariableReadNode that returns nil.
The initialization is problematic because it fixes the slot type already to objects. (And, I didn't get it to work properly because of that. The slot type turned out to be variable at runtime, which is bad.)

Now, we specialize reads for uninitialized slot types to return nil. This condition should always constant for a specific AST node, because I can't make up a situation for the current special handing of local and non local variables, where that would not work. We need to watch out when we unify local and non local variables again. Because, then, we can have loops on a variable, and the state of the variable might change during the loop, for the same AST node. But for the moment, that's not possible.

[TODO-REPORT: discuss this design, in case it survives]

Signed-off-by: Stefan Marr <git@stefan-marr.de>