Change access of outer self to use arguments array instead of frame slot.
The frame slot access cannot be resolved to a constant, because it comes from a SBlock object, and therefore cannot be optimized. While using the arguments array does not allow for specializing the access, it seems to get optimized better because it does not require access to a FrameSlot object.

Signed-off-by: Stefan Marr <git@stefan-marr.de>