Splitting enforced and unenforced trees into different methods [WIP]
Truffle/Graal do not particularly like the fact that both parts of the execution tree were in the same tree. By splitting it, we reduce the load on the optimizer etc. The idea is that we actually can determine statically whether to use the enforced or unenforced part, as part of the lookup.
This should help with performance, and make the trees in IGV more easy to read because it doesn't contain extraneous nodes.

- moved enforced/unenforced to SInvokable level
- also added a heuristic to ask Graal/Truffle to inline the meta operations
- added another helper SArguments.enforced(.)
- removed distinction between enforced and unenforced Invokable
- the dispatch nodes now typically get directly the CallTarget for later execution

Signed-off-by: Stefan Marr <git@stefan-marr.de>