Added basic Thread support
This is somewhat minimal support for using threads with SOM. Note, it is slightly different from previous SOM versions with threading. I tried to make the names as telling as possible. Furthermore, the set of added features was kept small and it should be possible to implement them on a wide varity of platforms.

Signed-off-by: Stefan Marr <git@stefan-marr.de>