Added first sketch of Platform module
This is just a very first idea based on what I understand a Newspeak implementation should do.

Note, at the moment, I used field assignments, which isn't really Newspeak. Not yet decided whether this is going to stay or not. Normally, Newspeak requires setter sends as far as I understand.

Signed-off-by: Stefan Marr <git@stefan-marr.de>