Simplify SArray and derive storage type directly from itself
In the compilation we got two guards now on every array operation, one for the type, and one for the storage. This is unfortunate... But, we actually don't need both since the storage is representing the same info.

Remove ArrayType enum, but keep it in the PartiallyEmptyArray to track population type.

Signed-off-by: Stefan Marr <git@stefan-marr.de>