Implement the handling of the return code in the VM
This is a first step to add VM-level actor support.
Don't want to expose the necessary knowledge at the language level.
SOMns is supposed to be a language where such information are only available if absolutely necessary. This includes for instance the number of actors in the system, based on which a VM could decide whether or not the main program is an event loop or just a process.

Signed-off-by: Stefan Marr <git@stefan-marr.de>