Rename all Class stuff to Mixin if it is not about Class objects
Need to get the terminology right now, otherwise it is going to be a horrible mess in the future. Since we introduced mixin handling, there is need for a clear conceptual distinction between classes and mixins.

In the Parser, we keep class in many cases, because it reveals more of the intention from the user perspective, and there is some overlap with what the runtime semantics is in terms of class slots etc.

Signed-off-by: Stefan Marr <git@stefan-marr.de>