Remove all class-class state from ClassFactory, and use separate instances
Having the class state in the ClassFactory was just not correct. Or at least really hard to get correct, because you need the context to know what needs to be accessed. Now, we essentially replicate the normal class/classClass split, and don't have issues if the classFactory/classGroup is supposed to be used in guards for instance

Signed-off-by: Stefan Marr <git@stefan-marr.de>