Optimize actor scheduling for many small messages
The old approach scheduled each actor message on the fork/join pool. This is a very expensive approach, because the scheduling overhead is high.

Now, we have a separate ExecAllMessages task, that is scheduled on the pool. This executor task will grab the mailbox of its actor, replace the mailbox on the actor with an empty one, and then start executing the mailbox. If it is done, it will go back to the actor, and attempt to take the current mailbox.

This approach is befenfitial for many small tasks. However, it probably reduces the fairness, because it will not yield the f/j pool, but continue executing until it sees an empty mailbox.

Furthermore, it might be slightly more expensive for the first message to be processed (this might be a minor).

EventualMessages are not RecursiveActions anymore, which should also make them slightly more lightweight.
Setting the current actor only needs to be done once per scheduling the executor.

Signed-off-by: Stefan Marr <git@stefan-marr.de>