Fixed Value related tests
Seems like the TransferObject implementation fixed a bug in the Value implementation as well: the outer/enclosing object was not checked to be a value on constructing a Value class.
Since we moved the exception now to the class construction, this should be solved. And it seems to be the right approach, because it would also most likely be the place where one wants and exception for a live environment.

Signed-off-by: Stefan Marr <git@stefan-marr.de>