Fix bug in handling STransferArrays and SObjectWithoutFields as TransferObjects
We did not yet cover the case that objects without fields can be transfer objects.
This is needed to be consistent with other types of objects. Perhaps sufficially it doesn't really make a lot of sense, except that they have their lexcial scope and are sequentially accessible, and to be consistent.

Also, we did not handle TransferArrays yet completely, which could lead to them being send as far references.

Signed-off-by: Stefan Marr <git@stefan-marr.de>