Prevent deadlock and work around race condition
The deadlock happens when processing promises and trying to wrap them. Here it can lead to a deadlock, because we traverse the graphs unordered/uncontrolled.

When breaking up this deadlock, we need to further avoid the race condition between the promise resolve and another thread trying to send messages to the promise. Because of the race, it could happen that messages from the same sender are arriving in wrong order. The resolve has resolved the promise, which becomes visible, but has not yet scheduled the old callbacks from the promise on the actor. But, the other actor already sees the resolved state of the promise, and enqueues a message directly on the actor, which then is out of order with its other messages.

- further refactorings and name clarifications

Signed-off-by: Stefan Marr <git@stefan-marr.de>