Prototype specialization to remove superflous setting of bit for primitive fields
Before this change, every write to a primitive field would also cause a write of the mark bit field.

Now, we only got a read, and a deopt, ideally.
We bring this rather stable behavior to the AST by having two cases, one that just checks the bit, and another one that also sets it. The later one should be rarely necessary.

TODO: remove code duplication, ideally convert to use DSL
Signed-off-by: Stefan Marr <git@stefan-marr.de>