Fix issue with overriding the primary factory method
The problem was that the primary factory is the last method entered into the table of factory methods, and so always overwrote the secondary factory methods, which seems wrong.

So, the solution is to take into account that the primary method is coming in last.

Also added a proper error in case there are two secondary factories with the same name.

Signed-off-by: Stefan Marr <git@stefan-marr.de>