Use step-after-root node to realize promise resolver breakpoint
This change improves the source attribution for the promise resolver breakpoint.
Previously, we use a haltNode (SuspendExecutionNode) to trigger a breakpoint in the promise resolution node with a source section that could not be attributed to the concrete method executed (because of polymorphic message sends).

Now, we stop execution after the rootNode finished executing, which means before exiting the concrete method that was activated.

Signed-off-by: Stefan Marr <git@stefan-marr.de>