Fix #72: Ensure locals of embedded method are set
Previously, the local variables of a block method that got inlined/embedded were not know in the method builder associate with it.
The problem was hidden by assignments adding these locals. So, only in cases where there were no previous assignments in the scope, lookup of a local that got embedded in the outer scope, would fail.

Signed-off-by: Stefan Marr <git@stefan-marr.de>