Compress dym test data
Before, the test data was included as CSV files, which caused huge diffs with little gain on every SOM change.

Now, the test data is compressed, which should keep the diffs clean. Diffs are only visible during testing now, which should be fine.

The test script now takes the `update` argument to update the test data.

Signed-off-by: Stefan Marr <git@stefan-marr.de>