Refactor AST inlining and use Variable objects
The goal of this change is to make the AST inlining more maintainable and
less complex. Furthermore, we want the Variable objects to encode more
information for debugging purposes.

Currently, AST inlining is essentially encoded in the inliners,
which operate on demand and add variable slots lazily.

With this rewrite, AST splitting and inlining is done based on the
tree of scopes. Thus, the correct construction of frame slots and
variable objects in localized in one place. And, it is done eagerly
based on upfront knowledge, so that bugs hopefully aren’t hidden,
which they were by the on-demand approach.

- moved visitor into inlining package
- populate lexical scopes eagerly with embedded info
  needed for the inlining at parse time, because we need the info directly
  once the parsing of the embedded scope is completed.
- variables are now equal if they refer to the same lexical entity
- when correcting ASTs after inlining/splitting, lookup variables in
  scope tree
- make sure scope tree updates embedded methods, need to be split, and
  inlined methods/blocks need to be removed

Signed-off-by: Stefan Marr <git@stefan-marr.de>