Add protocol overview visualization
- added new canvas to draw protocol.
- every actor receives header square
- every message received by an actor becomes a turn: circle below header
- messages are lines between turns

- nodes show name on hover
- when hidding actor both incoming and outgoing messages are hidden
- on click a node and the incoming message are highlighted
- added enlarge of turns, still need to add markers to serve as clickable objects
- added marker arrows and squares

- clicking turn in protocol overview now highlight method in code view

Source-Attribution and Origin:

This code was written by Sander <sander_l94@hotmail.com>
It is a squashed version, that removes all changes not related to the
frontend code. It also does not contain any code related to the
time-travelling features.

The code is taken from
https://github.com/MetaConc/SOMns/tree/time-travelling
in the version of May 20, 2017.
The last commit was fc1cb917c361035460c7e9028cb6af8121874f22
"fixed turn and actor id when clicking on popover"