Concept of MethodDeclaration can be known to debugger, fix their bp id lookup
The concept of a MethodDeclaration has been know to the debugger from early on.
I think, we could strip that out as well, but since it is a structural feature, I don’t really see the need.
It is not really a concurrency-related concept, thus, I won’t remove knowledge about it from the debugger.

Signed-off-by: Stefan Marr <git@stefan-marr.de>