Use ephemeral ports if standard ports are not available
Kompos tests fail on travis some times with an address already bound error.
To avoid it, try to use standard port, and if already in use, use an ephemeral port instead.
The web socket ports can be communicated via JSON and easily queried by the frontend.
The HTTP port needs to be parsed out of the STDOUT.

- update SOMns-deps to have latest WebSocket lib version
- in tests, parse stdout for ports
- remove duplicate code of old BinaryWebSocketHandler

Signed-off-by: Stefan Marr <git@stefan-marr.de>