Remove recording of time stamps
This is mostly to make my life easier.
They aren’t useful, at least, we haven’t really used them for something yet.
In the future, we might want them back, for some sort of performance monitoring, perhaps.
But, since they are rather expensive, it is unclear that’s a good strategy in the first place.

Signed-off-by: Stefan Marr <git@stefan-marr.de>