Added enclosing MethodScope to MixinScope
The MixinScopes representing an object literal need to have
a record of the corresponding enclosing MethodScope. These changes
added new constructors (for MixinScope) that enable MethodScopes to be
passed as an argument and also added helper methods to get either the
enclosing MethodScope (for object literals) or MixinScope (for classes)
from the MixinScope.