Add new way of writting stepping tests
The main benefit of this approach is that we can handle non-deterministic stepping results easily and don’t need to do it in the assertions manually.

This new stepping.ts introduces the data format for the tests, and it is currently used in actor.ts to generate tests.

The basic idea is that an initial set of breakpoints results in a single stop, from which we then can take separate steps, which each lead to n predetermined stops at given positions.

Note that the system uses aliases for the actors instead of ids which are fragile. It determines the activity id by matching it against the expected names used in the steps. So, it discovers the mapping at run time automatically. This works, as long as we introduce activities only gradually.

Signed-off-by: Stefan Marr <git@stefan-marr.de>