Abstracted class instantiation logic
A new instantation node designed for object literals has been added. The
caching behavior required for this node is similar to the existing
`ClassInstantiationNode`. These changes provide an abstract `InstantiationNode`
that implements the shared behavior. `ClassInstantiationNode` has been refactored
as an innersubclass of the `InstantiationNode` and `ObjectLiteralInstantiationNode` has been
added (also as an inner subclass).