New object literal node and adapted SClass
The object literal is created with a mixin definiton that corresponds to
this object's implicitally declared class. This new node is responsible
for first instantiating the implicit class and then the object from
that class.

A context field has also been added to the `SClass`, which is responsible for
holding the activation in which the object literal is created. Methods have
been added to `MixinDefintion` and `ClassInstantiationNode` to pass the the
`MaterializedFrame` as an argument to the instantiation.