Handle method/mixin scope uniformly
Until now, we kept mixin and method scopes separately, which does not directly represent the lexical structure and can cause confusion.

With this change, scope is managed as a single chain outwards, with methods and mixins on the same chain.

This change applies to LexicalScope classes as well as to the Builder classes.

- add ScopeBuilder superclass for MixinBuilder and MethodBuilder
- correctly determine whether scopes are immutable also for methods
- added Local.isMutable()
- InliningVisitor, handle case that there are no variables

Signed-off-by: Stefan Marr <git@stefan-marr.de>