Increase chance that objects are update with valid layouts
If multiple threads mutate objects from the same class, it can happen that the object layout gets invalidated while we do an update in another thread.

In the worst case, this can lead to stack overflows, because our interpreter tries repeatedly to find a valid object layout while creating dispatch chains, or interacting with the object layout in other nodes, which use recursive fallbacks.

To reduce the chance of this happening, I added a busy loop to wait that our lookup getLayoutForInstancesToUpdateObject() actually returns a valid layout. It might still be invalidate before use again, but hopefully this happens now less often, and prevents stack overflows.

Signed-off-by: Stefan Marr <git@stefan-marr.de>