Restructure RIRS.specialize to be race free
Carefully group all accesses to replacedBy and newReceiverNodeForOuterSend.
Technically, this could also be realized with volatile fields, but I wanted to have the more clear grouping with explicit synchronized, just to be sure.

The reordering makes sure all racy fields of the node are read and written at most once. This is to prevent any interference from other threads having problematic effects.

Signed-off-by: Stefan Marr <git@stefan-marr.de>