The Ast Builder [note]
This class is responsible for generating SOM AST and uses both the Source and Scope utilities (introduced in the previous commits). 

The current implementation provides only what is required to create an module that contains only a `#main:` method (that only returns zero). As the implementation grows it support be the creation of classes, objects, methods, literals, messages, and any other expressions we need to support the execution of Grace programs.

Finally, for convenience, I've added a new method for setting inheritance expressions easily to `MixinBuilder` and added module-based symbols to the `Symbols`utility.