The Json Tree Translator [note]
The `JsonTreeTranslator` is responsible for using the `AstBuilder` to assemble a SOM AST from a Grace AST (which is a parse tree expressed as JSON data structured obtained via Kernan). To perform the translation, the translator begins from the root node and walks through the tree using an "in-order" style (first touching the parent, then its children, then finishing the parent). At each step, the translator works with the builder to enable a recusrive-descent style of translation.

The translator assembles each method into a finished SOM invokable as it goes and, once finished, the class resulting from the translation of the module is ready to be instantiated directly by SOMns.