Translate class declarations and partial support for inherits [note]
Here I've introduced translation for Grace's class and inheritance expressions; at least to a first approximation. 

A Grace class is translated into (1) a SOM class and (2) a method that returns an instance of that class. The SOM class (1) has a munged name, whereas the method (2) has the expected Grace signature. The munging is necessary to avoid name conflicts. 

Finally, inherit expressions are translated into an implicit request for the named superclasses. The inherits expression supports any number of arguments, although these arguments must be literal for now. In particular, preserving scoping information is not trivial since the SOM class is NOT enclosed by the scope surrounding the inheritance expression. 

In the future, I will either change SOMns or the builder/translator to enable Moth to support more sophisticated inheritance expressions.