VM - new `SomStructuralType` class [note]
This class represents a structural type. We represent a type as no more than a list of expected method signatures. A SOM class that can respond to each of these signatures is considered to conform to the type. This class also defines constants that describe the names of Grace's built-in types.

Here I have also changed the `SClass.canUnderstand` method to check superclasses. This goes against conventional SOM, but supports a simple implementation for now.