Translator - object constructors are now understood [note]
Grace's object constructor nodes are now translated. An object constructor declares an anonymous class definition nested onto the enclosing object and then is translated into an expression that creates an instance of this class (via a SOM `Object Literal`).  Finally, the ScopeManager has been changed to return the MixinDefinition of assembled classes.