VM - new dispatch guards [note]
I've added a family of new type-checking dispatch guards: a no-op guard, a guard foor boolean, one for string, another numbers, and finally the full structural type-checking guard. One of these guards is choosen, for each argument, whenever a new `DispatchNode` is created (a dispatch node corresponds to a particular call site). The choice of guard depends on the expected type: if unknown then the no-op guard is used, if the expected type is built-in one of literal guards is used, or otherwise we use the full type checking guard.