Fix - attach source information to block[note]
During inlining, its possible that two locals of the same name might be added to the enclosing block's scope. To handle this, SOMns munges the variable name by adding the line and column number as a suffix to this name. Without the correct source information, the name munging adds the suffix `1:1`. With this change, the name munging works correctly since the correct line and column numbers are available.